implemented res.location
