+++ lib/response.js
@@ -588,9 +588,9 @@ res.cookie = function(name, val, options){
   return this;
 };
 
+
 /**
- * Redirect to the given `url` with optional response `status`
- * defaulting to 302.
+ * Set the location header to `url`.
  *
  * The given `url` can also be the name of a mapped url, for
  * example by default express supports "back" which redirects
@@ -598,46 +598,31 @@ res.cookie = function(name, val, options){
  *
  * Examples:
  *
- *    res.redirect('/foo/bar');
- *    res.redirect('http://example.com');
- *    res.redirect(301, 'http://example.com');
- *    res.redirect('http://example.com', 301);
- *    res.redirect('../login'); // /blog/post/1 -> /blog/login
+ *    res.location('/foo/bar').;
+ *    res.location('http://example.com');
+ *    res.location('../login'); // /blog/post/1 -> /blog/login
  *
  * Mounting:
  *
- *   When an application is mounted, and `res.redirect()`
+ *   When an application is mounted, and `res.location()`
  *   is given a path that does _not_ lead with "/". For 
  *   example suppose a "blog" app is mounted at "/blog",
- *   the following redirect would result in "/blog/login":
+ *   the following call to `res.location()` would result
+ *   in "/blog/login":
  *
- *      res.redirect('login');
+ *      res.location('login');
  *
- *   While the leading slash would result in a redirect to "/login":
+ *   While the leading slash would result in a location of "/login":
  *
- *      res.redirect('/login');
+ *      res.location('/login');
  *
  * @param {String} url
- * @param {Number} code
  * @api public
  */
 
-res.redirect = function(url){
+res.location = function(url){
   var app = this.app
-    , req = this.req
-    , head = 'HEAD' == req.method
-    , status = 302
-    , body;
-
-  // allow status / url
-  if (2 == arguments.length) {
-    if ('number' == typeof url) {
-      status = url;
-      url = arguments[1];
-    } else {
-      status = arguments[1];
-    }
-  }
+    , req = this.req;
 
   // setup redirect map
   var map = { back: req.get('Referrer') || '/' };
@@ -658,6 +643,52 @@ res.redirect = function(url){
     }
   }
 
+  // Respond
+  this.set('Location', url);
+  return this;
+};
+
+/**
+ * Redirect to the given `url` with optional response `status`
+ * defaulting to 302.
+ *
+ * The resulting `url` is determined by `res.location()`, so
+ * it will play nicely with mounted apps, relative paths,
+ * `"back"` etc.
+ *
+ * Examples:
+ *
+ *    res.redirect('/foo/bar');
+ *    res.redirect('http://example.com');
+ *    res.redirect(301, 'http://example.com');
+ *    res.redirect('http://example.com', 301);
+ *    res.redirect('../login'); // /blog/post/1 -> /blog/login
+ *
+ * @param {String} url
+ * @param {Number} code
+ * @api public
+ */
+
+res.redirect = function(url){
+  var app = this.app
+    , head = 'HEAD' == this.req.method
+    , status = 302
+    , body;
+
+  // allow status / url
+  if (2 == arguments.length) {
+    if ('number' == typeof url) {
+      status = url;
+      url = arguments[1];
+    } else {
+      status = arguments[1];
+    }
+  }
+
+  // Set location header
+  this.location(url);
+  url = this.get('Location');
+
   // Support text/{plain,html} by default
   this.format({
     text: function(){
@@ -676,7 +707,6 @@ res.redirect = function(url){
 
   // Respond
   this.statusCode = status;
-  this.set('Location', url);
   this.set('Content-Length', Buffer.byteLength(body));
   this.end(head ? null : body);
 };
